+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-B3NtbhqJ.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-BKQfzrSz.js";import{_ as a}from"./mermaid.core-Blh6wACb.js";import"./chunk-5VM5RSS4-ZfoM_k-7.js";import"./chunk-XXDRQBXY-D5WjKh1e.js";import"./chunk-POPQ4Y6H-DwfTZu5s.js";import"./chunk-F27PBJKO-CCFMUEHr.js";import"./channel-nawJXMWc.js";import"./index-Bs2FsUt9.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};